+++ src/main/java/de/aivot/GoverBackend/elements/models/RootElement.java
@@ -11,6 +11,7 @@ import de.aivot.GoverBackend.utils.MapUtils;
 
 import jakarta.annotation.Nonnull;
 import java.util.*;
+import java.util.function.Predicate;
 
 public class RootElement extends BaseElement {
     private String headline;
@@ -81,6 +82,24 @@ public class RootElement extends BaseElement {
         return children
                 .stream()
                 .map(s -> s.findChild(id))
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .findFirst();
+    }
+
+    public Optional<? extends BaseElement> findChild(@Nonnull Predicate<BaseElement> pred) {
+        Optional<StepElement> matchingStep = children
+                .stream()
+                .filter(pred)
+                .findFirst();
+
+        if (matchingStep.isPresent()) {
+            return matchingStep;
+        }
+
+        return children
+                .stream()
+                .map(s -> s.findChild(pred))
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .findFirst();